Move screens into seperate folder
