+++ screens/NavigateScreen.js
@@ -1,23 +1,23 @@
 import { Text, View, Image, SafeAreaView, TouchableOpacity } from 'react-native';
-import styles from './styles.js';
-import { navigate, speak } from './functions.js';
+import styles from '../styles.js';
+import { navigate, speak } from '../functions.js';
 
-export default function PreferencesScreen({ navigation }) {
-  message = "Now viewing: Preferences. Press bottom banner to return home. Press top right banner to repeat this message."
+export default function NavigateScreen({ navigation }) {
+  message = "Now viewing: Navigate. Press bottom banner to return home. Press top right banner to repeat this message."
   speak(message)
 
   return (
     <SafeAreaView style={styles.container}>
       {/* Title Banner */}
       <View style={styles.topBanner}>
-        <Text style={styles.titleText}>Preferences</Text>
+        <Text style={styles.titleText}>Navigate</Text>
 
         <TouchableOpacity style={styles.topRightBannerButton} onPress={() => speak(message)}>
-          <Image source={require('./assets/volume.png')}/>
+          <Image source={require('../assets/volume.png')}/>
         </TouchableOpacity>
 
         <TouchableOpacity style={styles.topLeftBannerButton} onPress={() => navigate(navigation, "Home")}>
-          <Image source={require('./assets/back.png')}/>
+          <Image source={require('../assets/back.png')}/>
         </TouchableOpacity>
       </View>
